requirements.txt
fastapi
uvicorn[standard]
openai
httpx
python-dotenv
flask
flask-cors
